Split drive_configuration and drive_configuration_updater
