+++ drive/CMakeLists.txt
@@ -28,7 +28,7 @@ add_executable(${PROJECT_NAME}
     platform-implementations/offset_servo.cpp
     implementations/drive_debug.cpp
     implementations/wheel_router.cpp
-    implementations/drive_configuration.cpp
+    implementations/drive_configuration_updater.cpp
     implementations/command_lerper.cpp
     implementations/mode_switcher.cpp
     implementations/tri_wheel_router.cpp
